refactor : rebook-members 모듈 하위에 members 코드 옮김
