+++ rebook-members/src/main/java/com/be/rebook/members/dto/UpdateDTO.java
@@ -1,4 +1,4 @@
-package com.be.rebook.domain.members.dto;
+package com.be.rebook.members.dto;
 
 import lombok.Getter;
 import lombok.Setter;
