+++ rebook-members/src/main/java/com/be/rebook/members/dto/UserinfoDTO.java
@@ -1,4 +1,4 @@
-package com.be.rebook.domain.members.dto;
+package com.be.rebook.members.dto;
 
 import lombok.Builder;
 import lombok.Getter;
