+++ rebook-members/src/main/java/com/be/rebook/members/entity/Majors.java
@@ -1,4 +1,4 @@
-package com.be.rebook.domain.members.entity;
+package com.be.rebook.members.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
