+++ settings.gradle
@@ -1,4 +1,5 @@
 rootProject.name = 'reBook'
 include 'rebook-members'
 include 'rebook-security'
+include 'untitled'
